+++ src/SeablastView.php
@@ -14,7 +14,6 @@ class SeablastView
 
     /** @var SeablastModel */
     private $model;
-
     /** @var array<mixed>|stdClass TODO: only Object */
     private $params;
 
@@ -100,6 +99,13 @@ class SeablastView
     {
         if (!$this->model->getConfiguration()->flag->status(SeablastConstant::FLAG_DEBUG_JSON)) {
             header('Content-Type: application/json; charset=utf-8'); //the flag turns-off this line
+        }
+        if (isset($this->params->status) && is_scalar($this->params->status)) {
+            // todo in_array((int),[allowed codes] to replace basic validation below
+            if ((int) $this->params->status < 100 || (int) $this->params->status > 599) {
+                throw new \Exception('Unknown HTTP code: ' . (int) $this->params->status);
+            }
+            http_response_code((int) $this->params->status); // Send the status code
         }
         $result = json_encode($data2json);
         Assert::string($result);