+++ CHANGELOG.md
@@ -17,6 +17,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### `Security` in case of vulnerabilities
 
+## [0.1.1] - 2024-01-12
+### Added
+- class Api\GenericRestApiJsonModel implements SeablastModelInterface
+- Exception\UploadException to translate (int) code to (string) message
+- class Api\ApiErrorModel to be directly used by application
+- JSON response uses `status` property as a response HTTP code.
+
+### Changed
+- improved SeablastMysqli error logging
+
+### Fixed
+- more ways to identify HTTPS from SERVER headers
+
 ## [0.1] - 2023-12-30
 ### Added
 - MVC architecture
@@ -56,5 +69,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
-[Unreleased]: https://github.com/WorkOfStan/seablast/compare/v0.1...HEAD
+[Unreleased]: https://github.com/WorkOfStan/seablast/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/WorkOfStan/seablast/compare/v0.1...v0.1.1
 [0.1]: https://github.com/WorkOfStan/seablast/releases/tag/v0.1